Move on_show_toolbars_change()
